ppc64: erase unnecessary segment info printing
In ppc64 loading, it will print the segment information. This is not
wanted since other Archs don't have this. People even think there's
something wrong. So erase it to make it be consistent with other Archs.

And if people really want to check these info, they can specify "-d"
option. They are printed in print_segments() too.

Signed-off-by: Baoquan He <bhe@redhat.com>
Signed-off-by: Simon Horman <horms@verge.net.au>