kexec/fs2dt: Fix sorting of device tree
Commit b02d735bf252 ('ppc64 kdump device_tree sort') added code to
sort device tree properties, but it had a few issues.

A compare routine needs to return -1 and 1. The special case for
sorting properties with unit addresses only returned 1 and ignored
the opposite case, which screwed up the sorting.

We were missing a few more things:

- Need to check both basenames are the same length
- Need to check both basenames match

I noticed this when looking at the NUMA topology after a kexec, and
it had shifted.

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Simon Horman <horms@verge.net.au>