kexec: remove ordering constraints from -x (--no-ifdown) and -y (--no-sync)
Currently, the two options must be the last ones to be honored;
otherwise, they can get silently ignored and both the manpage and help
text point it out.  This is error-prone and trivial to fix.  There
isn't much point in pointing something out in documentation when the
peculiarity can be removed with four lines of extra code.

Update option handling so that the two arguments are honored
regardless of their positions.

Signed-off-by: Tejun Heo <tj@kernel.org>
Signed-off-by: Simon Horman <horms@verge.net.au>