arm: Change setup_dtb_prop to create nodes by offset and node name
setup_dtb_prop looks like a generic function that can find and add any
nodes, which need not be top-level in the DT.  In practice though, the
function is only used for the top-level /chosen node, and it can't add
nodes for which the parent doesn't exist.

So far, so good - but for adding a new node to the DT, the parent offset
need be passed to fdt_add_subnode.  Currently in setup_dtb_prop the
parent offset is unknown, and instead a bogus error code is passed to
fdt_add_subnode.

Fix that by adding the parent offset as an extra function argument to
setup_dtb_prop, and change the handling of the /chosen node to operate
on a relative path plus (zero) offset instead of an absolute path. This
aligns setup_dtb_prop to the libfdt API, where functions commonly
operate with a parent offset plus child node name.

Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Signed-off-by: Simon Horman <horms@verge.net.au>