fs2dt: Fix for parsing/reading entries > 4k
The device-tree only does 4k reads at a time, so for entries greater
than that, the fs2dt code would fail, assuming a single read would get
the whole item.

Added helper function to continue reading as long as there was data and
no errors.

This was evident from:

/proc/device-tree/soc@ffe000000/fman@400000/fman-firmware/fsl,firmware

On my P4080 system, which is 29668 bytes.

Signed-off-by: Ben Collins <ben@cyphre.com>